Bayesian test for segregation distortion in tetraploids when genotypes are known.

This will run a Bayesian test using the genotypes of an F1 population of tetraploids for the null of Mendelian segregation (accounting for double reduction and preferential pairing) against the alternative of segregation distortion. This is when the genotypes are assumed known.

bayes_men_g4(
  x,
  g1,
  g2,
  drbound = 1/6,
  shape1 = 5/9,
  shape2 = 10/9,
  ts1 = 1,
  ts2 = 1,
  pp = TRUE,
  dr = TRUE,
  xi1 = 1/3,
  xi2 = 1/3,
  alpha = 0,
  ...
)

Arguments

x

A vector of genotype counts. x[i] is the number of offspring with genotype i-1.

g1

The genotype of parent 1.

g2

The genotype of parent 2.

drbound

The maximum rate of double reduction. A default of 1/6 is provided, which is the rate under the complete equational segregation model of meiosis.

shape1

The shape 1 parameter for the beta prior over the preferential pairing parameter.

shape2

The shape 2 parameter for the beta prior over the preferential pairing parameter.

ts1

The shape 1 parameter for the beta prior over the rate of quadrivalent formation.

ts2

The shape 2 parameter for the beta prior over the rate of quadrivalent formation.

pp

A logical. Should we account for preferential pairing (TRUE) or not (FALSE)?

dr

A logical. Should we account for double reduction (TRUE) or not (FALSE)?

xi1

If pp = FALSE, then xi is the known rate of preferential pairing (assumed fixed) of parent 1.

xi2

If pp = FALSE, then xi is the known rate of preferential pairing (assumed fixed) of parent 2.

alpha

If dr = FALSE, then alpha is the known rate of double reduction (assumed fixed).

...

Additional arguments to pass to Stan.

Value

A list with the following elements

lbf

The log Bayes Factor.

m0

Marginal likelihood under null.

ma

Marginal likelihood under alternative.

alpha

The posterior mean double reduction rate.

xi1

The posterior mean preferential pairing parameter of parent 1.

xi2

The posterior mean preferential pairing parameter of parent 2.

Impossible genotypes

Some offspring genotype combinations are impossible given the parental gentoypes. If these impossible genotypes combinations show up, we return a log-BF of -Inf and missing values for all other return items. The impossible genotypes are (when double reduction is allowed):

g1 == 0 && g2 == 0

Only offspring genotypes of 0 are possible.

g1 == 4 && g2 == 4

Only offspring genotypes of 4 are possible.

g1 == 0 && g2 == 4 || g1 == 4 && g2 == 0

Only offspring genotypes of 2 are possible.

g1 == 0 && g2 %in% c(1, 2, 3) || g1 %in% c(1, 2, 3) && g2 == 0

Only offspring genotypes of 0, 1, and 2 are possible.

g1 == 4 && g2 %in% c(1, 2, 3) || g1 %in% c(1, 2, 3) && g2 == 4

Only offspring genotypes of 2, 3, and 4 are possible.

When double reduction is 0, the list of impossible genotypes is more complicated.

Unidentified parameters

When g1 = 2 or g2 = 2 (or both), the model is not identified and those estimates (alpha, xi1, and xi2) are meaningless. Do NOT interpret them.

The estimate of alpha (double reduction rate) IS identified as long as at least one parent is simplex, and no parent is duplex. However, the estimates of the double reduction rate have extremely high variance.

Author

David Gerard

Examples

set.seed(1)
gf <- offspring_gf_2(alpha = 1/12, xi1 = 0.1, xi2 = 0.5, p1 = 2, p2 = 2)
x <- offspring_geno(gf = gf, n = 20)
bayes_men_g4(x = x, g1 = 2, g2 = 2, chains = 1, iter = 1000)
#> 
#> SAMPLING FOR MODEL 'marg_dr_pp_g4' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.000161 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 1.61 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 1000 [  0%]  (Warmup)
#> Chain 1: Iteration: 100 / 1000 [ 10%]  (Warmup)
#> Chain 1: Iteration: 200 / 1000 [ 20%]  (Warmup)
#> Chain 1: Iteration: 300 / 1000 [ 30%]  (Warmup)
#> Chain 1: Iteration: 400 / 1000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 500 / 1000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 501 / 1000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.108 seconds (Warm-up)
#> Chain 1:                0.095 seconds (Sampling)
#> Chain 1:                0.203 seconds (Total)
#> Chain 1: 
#> $lbf
#> [1] 3.220833
#> 
#> $m0
#> [1] -6.050226
#> 
#> $ma
#> [1] -9.271059
#> 
#> $alpha
#> [1] 0.04813376
#> 
#> $xi1
#> [1] 0.2970489
#> 
#> $xi2
#> [1] 0.3035922
#>