Bayesian test for segregation distortion in tetraploids using genotype likelihoods

This will run a Bayesian ratio test using the genotypes of an F1 population of tetraploids for the null of Mendelian segregation (accounting for double reduction and preferential pairing) against the alternative of segregation distortion. This is when genotype uncertainty is accounted for through genotype likelihoods.

bayes_men_gl4(
  gl,
  g1 = rep(-log(5), 5),
  g2 = rep(-log(5), 5),
  drbound = 1/6,
  shape1 = 5/9,
  shape2 = 10/9,
  ts1 = 1,
  ts2 = 1,
  pp = TRUE,
  dr = TRUE,
  alpha = 0,
  xi1 = 1/3,
  xi2 = 1/3,
  parent_method = c("est", "marg"),
  ...
)

Arguments

gl

The genotype log-likelihoods. The rows index the individuals and the columns index the genotypes.

g1

Either parent 1's genotype, or parent 1's genotype log-likelihoods.

g2

Either parent 2's genotype, or parent 2's genotype log-likelihoods.

drbound

The maximum rate of double reduction. A default of 1/6 is provided, which is the rate under the complete equational segregation model of meiosis.

shape1

The shape 1 parameter for the beta prior over the preferential pairing parameter.

shape2

The shape 2 parameter for the beta prior over the preferential pairing parameter.

ts1

The shape 1 parameter for the beta prior over the rate of quadrivalent formation.

ts2

The shape 2 parameter for the beta prior over the rate of quadrivalent formation.

pp

A logical. Should we account for preferential pairing (TRUE) or not (FALSE)?

dr

A logical. Should we account for double reduction (TRUE) or not (FALSE)?

alpha

If dr = FALSE, then alpha is the known rate of double reduction (assumed fixed).

xi1

If pp = FALSE, then xi is the known rate of preferential pairing (assumed fixed) of parent 1.

xi2

If pp = FALSE, then xi is the known rate of preferential pairing (assumed fixed) of parent 2.

parent_method

If unknown, should we estimate the parent genotypes by maximum likelihood ("est") or marginalize over them ("marg")?

...

Additional arguments to pass to Stan.

Value

A list with the following elements

lbf

The log Bayes Factor.

m0

Marginal likelihood under null.

ma

Marginal likelihood under alternative.

alpha

The posterior mean double reduction rate.

xi1

The posterior mean preferential pairing parameter of parent 1.

xi2

The posterior mean preferential pairing parameter of parent 2.

Unidentified parameters

When g1 = 2 or g2 = 2 (or both), the model is not identified and those estimates (alpha, xi1, and xi2) are meaningless. Do NOT interpret them.

The estimate of alpha (double reduction rate) IS identified as long as at least one parent is simplex, and no parent is duplex. However, the estimates of the double reduction rate have extremely high variance.

Author

David Gerard

Examples

if (FALSE) {
## null sims
set.seed(10)
g1 <- 1
g2 <- 2
gl <- simf1gl(n = 100, g1 = g1, g2 = g2)
bayes_men_gl4(gl = gl, g1 = g1, g2 = g2, chains = 1, iter = 1000)

## genotypes of parents not known
bayes_men_gl4(gl = gl, chains = 1, iter = 1000)

## alt sims
gl <- hwep::simgl(nvec = rep(5, 5), rdepth = 10)
bayes_men_gl4(gl = gl, g1 = g1, g2 = g2, chains = 1, iter = 1000)

bayes_men_gl4(gl = gl, chains = 1, iter = 1000)
}