menbayes: Tests for Mendelian Segregation in Polyploids

Provides a suite of tests for segregation distortion in F1 polyploid populations (for now, just tetraploids). This is under different assumptions of meiosis. This material is based upon work supported by the National Science Foundation under Grant No. 2132247. The opinions, findings, and conclusions or recommendations expressed are those of the author and do not necessarily reflect the views of the National Science Foundation.

See also

Author

Maintainer: David Gerard gerard.1787@gmail.com (ORCID)

Authors:

Mira Thakkar mt7562a@american.edu

Other contributors:

NSF DBI 2132247 (https://www.nsf.gov/awardsearch/showAward?AWD_ID=2132247) [funder]