Run segregation distortion tests as implemented in the polymapR package.

The polymapR package tests for segregation distortion by iterating through all possible forms of disomic or polysomic inheritance from either parent, tests for concordance of the offspring genotypes using a chi-squared test, and returns the largest p-value. It sometimes chooses a different p-value based on other heuristics. They also sometimes return NA. When type = "menbayes", we only look at patterns of the given parent genotypes, choosing the largest p-value. When type = "polymapR", we return what they use via their heuristics.

polymapr_test(x, g1, g2, type = c("menbayes", "polymapR"))

Arguments

x

Either a vector of genotype counts, or a matrix of genotype posteriors where the rows index the individuals and the columns index the genotypes.

g1

Parent 1's genotype.

g2

Parent 2's genotype.

type

Either my implementation which approximates that of polymapR ("menbayes") or the implementation through polymapR ("polymapR"). Note that polymapR needs to be installed for type = "polymapR".

Value

A list with the following elements:

p_value

The p-value of the test.

bestfit

The best fit model, using the same notation as in checkF1().

frq_invalid

The frequency of invalid genotypes.

See also

Author

David Gerard

Examples

g1 <- 1
g2 <- 2
gf <- offspring_gf_2(alpha = 0, xi1 = 1/3, xi2 = 1/3, p1 = g1, p2 = g2)
x <- offspring_geno(gf = gf, n = 100)
polymapr_test(x = x, g1 = g1, g2 = g2, type = "menbayes")
#> $p_value
#> [1] 0.3978336
#> 
#> $best_fit
#> [1] "1551_0"
#> 
#> $frq_invalid
#> [1] 0
#> 
#> $p_invalid
#> [1] 1
#>